+++ ea/AdjacencyMatrix.h
@@ -5,11 +5,11 @@
 class AdjacencyMatrix
 {
 
-    int** matrix = nullptr;
+    float** matrix = nullptr;
 
     void allocateMemory();
-    void insertValues(const int* data_);
-    void copyValues(int** dataFromOrigin_);
+    void insertValues(const float* data_);
+    void copyValues(float** dataFromOrigin_);
 
     void deallocateMemory();
 
@@ -17,16 +17,16 @@ public:
 
     const size_t citiesNumber;
 
-    AdjacencyMatrix(size_t citiesNumber_, const int* data_);
+    AdjacencyMatrix(size_t citiesNumber_, const float* data_);
     AdjacencyMatrix(const AdjacencyMatrix& origin_);
     ~AdjacencyMatrix();
 
-    int** getMatrix() const
+    float** getMatrix() const
     {
         return this->matrix;
     }
 
-    int getValue(size_t source_, size_t destination_) const
+    float getValue(size_t source_, size_t destination_) const
     {
         return this->matrix[source_][destination_];
     }
